+++ mRemoteV1/Config/Connections/ConnectionsLoader.cs
@@ -15,7 +15,6 @@ using mRemoteNG.Tree;
 using mRemoteNG.Connection;
 using mRemoteNG.Container;
 using mRemoteNG.Connection.Protocol;
-using mRemoteNG.Images;
 using mRemoteNG.UI.Forms;
 using mRemoteNG.Tree.Root;
 using mRemoteNG.UI.TaskDialog;
