+++ mRemoteV1/Config/Import/RemoteDesktopConnection.cs
@@ -3,10 +3,9 @@ using System.Windows.Forms;
 using System.IO;
 using mRemoteNG.App;
 using mRemoteNG.Connection.Protocol.RDP;
-using mRemoteNG.Images;
 using mRemoteNG.Container;
 using mRemoteNG.Connection;
-
+using mRemoteNG.Tree;
 
 namespace mRemoteNG.Config.Import
 {
